Changed: EVENT Limit for SERVER to get around issues with level coop_iceStation
This should not have any Negative side effects, except consuming more RAM.
This appears to be NOT linked to the CLIENT
In general we should look at some point into this thing.
This appears to be a issue with the info_pathnodes.
I do not yet know why they are even handled in events, a path file is build once and updated if nodes change.
If we have nothing else to do we could look into this...